fix tracing around cycle boundaries
previously we would log a threadswitch whenever the current tid changes,
or if the cycle just started and the previous job tid isn't zero.

now we properly log a threadswitch if the _next_ job tid isn't zero.
we only set cycle_started to false if we run a non-tid0 task.

we still log on every tidswitch boundary if (not cycle_started).